refactor: move get_sync_queue function to a dedicated queues module
