+++ frappe_gmail_thread/tasks/sync.py
@@ -1,7 +1,7 @@
 import frappe
 from frappe.utils.background_jobs import is_job_enqueued
 
-from frappe_gmail_thread.utils.helpers import get_sync_queue
+from frappe_gmail_thread.utils.queues import get_sync_queue
 
 
 def sync_emails():
